+++ test/models/Tweet.spec.js
@@ -1,96 +1,122 @@
-const chai = require('chai')
-const sinon = require('sinon')
-chai.use(require('sinon-chai'))
-
-const { expect } = require('chai')
-const {
-  sequelize,
-  dataTypes,
-  checkModelName,
-  checkPropertyExists
-} = require('sequelize-test-helpers')
-
+var chai = require('chai')
+var request = require('supertest')
+var sinon = require('sinon')
+var app = require('../../app')
+var helpers = require('../../_helpers');
+var should = chai.should();
+var expect = chai.expect;
 const db = require('../../models')
-const TweetModel = require('../../models/tweet')
+const passport = require('../../config/passport')
 
-describe('# Tweet Model', () => {
-  // 使用寫好的 Tweet Model
-  const Tweet = TweetModel(sequelize, dataTypes)
-  // 創建 tweet instance
-  const like = new Tweet()
-  // 檢查 Model name
-  checkModelName(Tweet)('Tweet')
+describe('# tweet requests', () => {
 
-  // 檢查 tweet 是否有 ___ 屬性(由於希望學員可以彈性命名 model 欄位，因此這邊留空)
-  context('properties', () => {
-    ;[
-    ].forEach(checkPropertyExists(like))
-  })
+  context('# POST ', () => {
 
-  // 檢查 tweet 的關聯是否正確
-  context('associations', () => {
-    const Reply = 'Reply'
-    const Like = 'Like'
-    const User = 'User'
-    before(() => {
-      // 將 Tweet model 對 User, Tweet, Reply 做關聯(呼叫 associate)
-      Tweet.associate({ Reply })
-      Tweet.associate({ Like })
-      Tweet.associate({ User })
-    })
-
-    it('should have many replies', (done) => {
-      // 檢查是否有呼叫 hasMany(Reply)
-      expect(Tweet.hasMany).to.have.been.calledWith(Reply)
-      done()
-    })
-    it('should have many likes', (done) => {
-      // 檢查是否有呼叫 hasMany(Like)
-      expect(Tweet.hasMany).to.have.been.calledWith(Like)
-      done()
-    })
-    it('should belong to user', (done) => {
-      // 檢查是否有呼叫 belongsTo(User)
-      expect(Tweet.belongsTo).to.have.been.calledWith(User)
-      done()
-    })
-  })
-
-  // 檢查 model 的新增、修改、刪除、更新
-  context('action', () => {
-    let data = null
-
-    // 檢查 db.Tweet 是否真的可以新增一筆資料
-    it('create', (done) => {
-      db.Tweet.create({ UserId: 1, description: 'hi' }).then((tweet) => {
-        data = tweet
-        done()
+    describe('POST /api/tweets', () => {
+      before(async() => {
+        // 清除測試資料庫資料
+        await db.User.destroy({where: {},truncate: true})
+        await db.Tweet.destroy({where: {},truncate: true})
+        // 模擬登入資料
+        const rootUser = await db.User.create({name: 'root'});this.authenticate =  sinon.stub(passport,"authenticate").callsFake((strategy, options, callback) => {            
+          callback(null, {...rootUser}, null);
+          return (req,res,next)=>{};
+        });
+        this.getUser = sinon.stub(
+            helpers, 'getUser'
+        ).returns({id: 1, Followings: []});
+        // 在測試資料庫中，新增 mock 資料
+        await db.User.create({account: 'User1', name: 'User1', email: 'User1', password: 'User1'})
       })
-    })
-    // 檢查 db.Tweet 是否真的可以讀取一筆資料
-    it('read', (done) => {
-      db.Tweet.findByPk(data.id).then((tweet) => {
-        expect(data.id).to.be.equal(tweet.id)
-        done()
+
+      // 新增推文 - POST /tweets
+      it(' - successfully', (done) => {
+        request(app)
+          .post('/api/tweets')
+          .send('description=description')
+          .set('Accept', 'application/json')
+          .expect(200)
+          .end(function(err, res) {
+            if (err) return done(err);
+            // 檢查是否有回傳正確資料
+            db.Tweet.findByPk(1).then(tweet => {
+              tweet.description.should.equal('description');
+              tweet.UserId.should.equal(1);
+              return done();
+            })
+          })
+      });
+
+      after(async () => {
+        this.authenticate.restore();
+        this.getUser.restore();
+        await db.User.destroy({where: {},truncate: true})
+        await db.Tweet.destroy({where: {},truncate: true})
       })
-    })
-    // 檢查 db.Tweet 是否真的可以更新一筆資料
-    it('update', (done) => {
-      db.Tweet.update({}, { where: { id: data.id } }).then(() => {
-        db.Tweet.findByPk(data.id).then((tweet) => {
-          expect(data.updatedAt).to.be.not.equal(tweet.updatedAt)
-          done()
-        })
+
+    });
+
+  });
+
+  context('# GET ', () => {
+
+    describe('GET /api/tweets', () => {
+      before(async() => {
+        // 清除測試資料庫資料
+        await db.User.destroy({where: {},truncate: true})
+        await db.Tweet.destroy({where: {},truncate: true})
+        // 模擬登入資料
+        const rootUser = await db.User.create({name: 'root'});this.authenticate =  sinon.stub(passport,"authenticate").callsFake((strategy, options, callback) => {            
+          callback(null, {...rootUser}, null);
+          return (req,res,next)=>{};
+        });
+        this.getUser = sinon.stub(
+            helpers, 'getUser'
+        ).returns({id: 1, Followings: []});
+        // 在測試資料庫中，新增 mock 資料
+        await db.User.create({account: 'User1', name: 'User1', email: 'User1', password: 'User1'})
+        await db.Tweet.create({UserId: 1, description: 'User1 的 Tweet1'})
       })
-    })
-    // 檢查 db.Tweet 是否真的可以刪除一筆資料
-    it('delete', (done) => {
-      db.Tweet.destroy({ where: { id: data.id } }).then(() => {
-        db.Tweet.findByPk(data.id).then((tweet) => {
-          expect(tweet).to.be.equal(null)
-          done()
-        })
+
+      // GET /tweets - 所有推文，包括推文作者
+      it(' - successfully', (done) => {
+        request(app)
+          .get('/api/tweets')
+          .set('Accept', 'application/json')
+          .expect(200)
+          .end(function(err, res) {
+            if (err) return done(err);
+            expect(res.body).to.be.an('array');
+            // 檢查是否回傳資料有 User1 的 Tweet1
+            res.body[0].description.should.equal('User1 的 Tweet1');
+            return done();
+          })
+      });
+
+      // GET /tweets/:tweet_id - 一筆推文與回覆
+      it(' - successfully', (done) => {
+        request(app)
+          .get('/api/tweets/1')
+          .set('Accept', 'application/json')
+          .expect(200)
+          .end(function(err, res) {
+            if (err) return done(err);
+            expect(res.body).to.be.an('object');
+            // 檢查是否回傳資料有 User1 的 Tweet1 
+            res.body.description.should.equal('User1 的 Tweet1');
+            return done();
+          })
+      });
+
+      after(async () => {
+        this.authenticate.restore();
+        this.getUser.restore();
+        await db.User.destroy({where: {},truncate: true})
+        await db.Tweet.destroy({where: {},truncate: true})
       })
-    })
-  })
-})
+
+    });
+
+  });
+
+});
